4327 - Move and deprecate migration-steps updateDependencies
