+++ src/server/controller/cycleData/tableData/index.ts
@@ -1,9 +1,11 @@
 import { getAggregatedTableData } from './getAggregatedTableData'
 import { massiveInsert } from './massiveInsert'
 import { refreshViews } from './refreshViews'
+import { updateTableDataDependencies } from './updateTableDataDependencies'
 
 export const TableData = {
   getAggregatedTableData,
-  refreshViews,
   massiveInsert,
+  refreshViews,
+  updateTableDataDependencies,
 }
