+++ src/server/controller/cycleData/tableData/massiveInsert.ts
@@ -1,16 +1,15 @@
 import { CountryIso } from 'meta/area'
 import { ActivityLogMessage, Assessment, Cycle, Node, RowCache, RowCaches } from 'meta/assessment'
-import { NodeUpdate, NodeUpdates } from 'meta/data'
+import { NodeUpdate } from 'meta/data'
 import { User } from 'meta/user'
 
-import { ContextFactory } from 'server/controller/cycleData/updateDependencies/context'
-import { updateCalculationDependencies } from 'server/controller/cycleData/updateDependencies/updateCalculationDependencies'
-import { BaseProtocol, DB } from 'server/db'
+import { DB } from 'server/db'
 import { NodeDb, NodeRepository } from 'server/repository/assessmentCycle/node'
 import { ActivityLogDb, ActivityLogRepository } from 'server/repository/public/activityLog'
 import { DataRedisRepository } from 'server/repository/redis/data'
 import { RowRedisRepository } from 'server/repository/redis/row'
-import { Logger } from 'server/utils/logger'
+
+import { updateTableDataDependencies } from './updateTableDataDependencies'
 
 type Props = {
   assessment: Assessment
@@ -18,51 +17,6 @@ type Props = {
   countryNodes: { [key in CountryIso]?: Array<NodeUpdate> }
   nodes: Array<NodeDb>
   user: User
-}
-
-export const _updateDependencies = async (
-  props: Pick<Props, 'assessment' | 'cycle' | 'countryNodes'>,
-  client: BaseProtocol = DB
-): Promise<void> => {
-  const { assessment, cycle, countryNodes } = props
-  const allNodesDb: Array<NodeDb> = []
-  const allNodes: Array<{ nodes: Record<string, Array<NodeUpdate>>; countryIso: CountryIso }> = []
-
-  const _execJob = async (props: { countryIso: CountryIso; nodes: Array<NodeUpdate> }): Promise<void> => {
-    const { countryIso, nodes } = props
-    const assessmentName = assessment.props.name
-    const cycleName = cycle.name
-
-    const nodeUpdates: NodeUpdates = { assessmentName, cycleName, countryIso, nodes }
-    const contextProps = { assessment, cycle, nodeUpdates, includeSourceNodes: false }
-    const context = await ContextFactory.newInstance(contextProps, client)
-    const result = updateCalculationDependencies({ context, jobId: `massive_insert-${Date.now()}` })
-
-    if (result.nodesDb.length > 0) {
-      allNodesDb.push(...result.nodesDb)
-      allNodes.push({ nodes: result.nodes, countryIso })
-    }
-  }
-
-  await Promise.all(
-    Object.entries(countryNodes).map(([countryIso, nodes]: [CountryIso, Array<NodeUpdate>]) =>
-      _execJob({ countryIso, nodes })
-    )
-  )
-
-  try {
-    if (allNodesDb.length > 0) {
-      await NodeRepository.massiveInsert({ assessment, cycle, nodes: allNodesDb }, client)
-      await Promise.all(
-        allNodes.map(async ({ countryIso, nodes }) => {
-          await DataRedisRepository.updateNodes({ assessment, cycle, countryIso, nodes })
-        })
-      )
-    }
-  } catch (e) {
-    Logger.error('Persisting nodes failed')
-    Logger.error(e)
-  }
 }
 
 export const massiveInsert = async (props: Props): Promise<void> => {
@@ -119,6 +73,6 @@ export const massiveInsert = async (props: Props): Promise<void> => {
     )
 
     // Update Deps
-    await _updateDependencies({ assessment, cycle, countryNodes })
+    await updateTableDataDependencies({ assessment, cycle, countryNodes })
   })
 }
